+++ src/kernel/cp_install.py
@@ -1,18 +1,67 @@
 #!/usr/bin/python3
+#
+# CPort - a new port manager for Calmira Linux
+# Copyright (C) 2021, 2022 Michail Krasnov <linuxoid85@gmail.com>
+#
+# cp_install.py
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+#
+# Project Page: http://github.com/Linuxoid85/cport
+# Michail Krasnov (aka Linuxoid85) linuxoid85@gmail.com
+#
 
 import os
 import wget
 import time
 import tarfile
-import subprocess
 import cp_default as cdf
 import cp_info    as cpI
+
+try:
+    import sqlite3
+
+    db     = cdf.DB
+    conn   = sqlite3.connect(db)
+    cursor = conn.cursor()
+except ImportError:
+    cdf.msg().error_trace(
+        f"It is not possible to use the cp_install kernel module: the" \
+        " 'sqlite3' package doesn't installed",
+        "cp_install kernel module"
+    )
+    raise cdf.KernelModuleImportError
 
 PORT_DIR  = cdf.PORT_DIR
 CACHE_DIR = cdf.CACHE_DIR
 CACHE_DATA_DIR = cdf.CACHE_DATA_DIR
 
+class check():
+
+    def filesystem(self, port) -> bool:
+        # TODO: перенести в cp_default
+        port = PORT_DIR + port
+        return os.path.isdir(port)
+
 class prepare():
+
+    def check_in_db(self, port):
+        data = "SELECT * FROM installed_ports WHERE name = (?)"
+        _db = cursor.execute(data, (port,))
+
+        return _db.fetchone() is None
     
     def download(self, port):
         url  = cpI.get().port_info(port, "port_management", "url")
@@ -79,11 +128,25 @@ class build(prepare):
         port_config = PORT_DIR + port + "/config.ini"
         port_install = PORT_DIR + port + "/install"
 
-        run = subprocess.run(port_install, shell=True)
+        run = os.system(port_install)
 
-        return run.returncode
+        return run
 
     def add_in_db(self, port):
         """
         TODO: add a code
         """
+
+        data = ""
+
+        try:
+            cursor.execute(data)
+            conn.commit()
+
+            return True
+        except sqlite3.DatabaseError as error:
+            cdf.msg().error_trace(
+                error,
+                f"cpi.build.add_in_db({port})"
+            )
+            return False